Add workaround in Shader.cpp for rendering without clip control
When you specify -DRIO_NO_CLIP_CONTROL, or use the new "no_clip_control"
Makefile target, this will now use a workaround I added to Shader.cpp
that will make the face look the right side up, rather than upside down.

I decided to flip texCoord for now, which didn't require modifying the
shader itself, but... I feel like the original texCoord value that is
being modified, it comes from FFL, so it would be more ideal to just
patch it there, which would mean any program using it wouldn't have to
accomodate its shader.

However, this workaround seems to work well, everything is right side up
in RenderDoc - the xlu mask AND the faceline texture (wrinkle and
makeup).

This will be necessary to run this on macOS, which it still doesn't as
of this commit, just throws an error at some glDrawElements call...