Use CPU SSAA scaling & move funcs in test server, others
* Add "scale" parameter to Go server, only 2 for now, will manually
request a higher resolution and downscale in Go. Rudimentary SSAA in
CPU.
- Actually, it isn't slow because it's copying and scaling on the CPU,
it's slow because there's more pixels that glReadPixels has to copy in
the renderer.

* Remove mouth/lip color conversion in DataUtils.cpp, upper lip common
colors have been added to my FFL fork. Now just to add the Switch glass
types, which FFL doesn't have.

* Make texture resolution high on low resolutions in the server.

* Experiment with FXAA. However, it didn't solve any of my problems with
GPU scaling, the fuzzy borders. So, I disabled it and SSAA.
* Experiment with no clip control, which probably still won't work on
this commit.