Update submodules
These new versions should enable building on MSVC. You still need to
make a vcxproj, I didn't include mine though since I'm not sure how to
elegantly include GLFW, GLFW, and zlib.

If you want to make your own, you should be able to just: define
includes, define defs, and include GLFW, GLEW, and zlib as includes and
libraries, and finally set it to emit the build to the root of the repo
so it can properly load assets, and then HOPEFULLY it should work.