Add initial headwear implementation.
* Add and set mpHeadwear property on Model.
* Add headwearIndex, headwearColor (defaults both -1) to
RenderRequest/server
* Add HeadwearModel class for loading and drawing headwear models.
* HeadwearList contains an std::vector of valid headwear models, a
method to load them from CSV and a method to get a HeadwearItem,
containing properties of a headwear model, by its index.

Sssssooooooooo the idea behind this is that all of the headwear models'
rmdl names, types, IDs are in this headwear.csv (in fs/content/), which
also has a path to its texture, both of which are opened and loaded in
every single time the headwear model is created. If you don't have
headwear.csv (like not included here) then mHeadwearAvailable will be
false and you will simply not be able to use headgear. ig i will give
kat his headwears and illllll ad the list for THAT  in
headwear.sample.csv