Add proper body scaling/skinning implementation, for Wii U/Switch body models
This implementation builds on the previous commit that added support for
skinning to all of the shaders, and here, a tiny bit of a hack is done
to still provide one static pose for the bodies but use skinning to
provide body scaling.

The matrices for each bone are in CSVs, where body_models.csv is also a
list of available body models defining which ones  have skeletons
(miitomo and 3ds do not right now). The bone index uniform is in
texCoord since the body models are still rmdl.

In BodyModel.cpp, UpdateScale and CalculateWorldMatrix are functionally
similar to nn::mii functions. The Wii U and Switch body models have the
exact same bones so it works for both but this will need adjustments to
work with miitomo and 3ds. However, that doesn't mean the basic logic
isn't the same, pretty much just UpdateScale() that needs modification.

In RootTask.cpp some modifications have been made to camera views for
face and all_body to match Wii U Mii Maker a bit closer (still guessing
by hand), and an adjustment was made to the camera position toooooo make
it kind of like the inverse of the model matrix? With this change the
camera view (and body scaling/switch shader ofc) should match
nn::mii::VariableIconBody perfectly one-to-one.

More details I didn't mention include the script I made to get CSV
matrices from ResourceVariableIconBodyHeader or glTFs, skl_root
adjustment on the Switch body to bring it a bit closer to Wii U body,
and how the bone indices for both body models need to have 3 added to
them.